Fix the title up on this, removed the date and format it to match the RELEASE NOTES.
Adaptec controllers are now 154x and 174x series, no more reference
to specific models.  Revamp the CSI hard disk controller section in
general to be more user readable.

Add the fact that the Mitsumi CDROM controller and drive are now
supported.

Add a note that the Intel 82501 serial chip is NOT supported.

Floppy controller is fd0, not fdc0, same for wd disk controller.