Remove all references to tickadj(8) from rc, rc.conf and rc.conf.5.
Disable building tickadj(8) by removing util from SUBDIR in the xntpd
Makefile. Note that the sources are still there and tickadj can still
be built and installed by doing:

# cd /usr/src/usr.sbin/xntpd/util
# make all install

There are enough references to tickadj in e.g. the xntpd documentation
(not to mention the sysctl variables it uses etc.) that I don't feel
up to implementing the final solution right now.

Kinda-approved-by:	phk