Changes have been made to support a concept of VAR_FIXIT_TTY being standard or serial.  This change needs to be done to the entire system that depends on this.  This way we don't have some code using OnVTY checks and other doing         strcmp(variable_get(VAR_FIXIT_TTY), "standard") == 0 checks.  Also we need to set VAR_FIXIT_TTY to "serial" if we come up on a serial console.
Also fixed a dialog problem in that dialog was used when dialog was
disabled causing some troubles such as not letting the cursor keys
work when exiting the fixit mode on media (ie. not the fixit shell but
for example fixit on a floppy).

Submitted by:	Doug Ambrisko <ambrisko@whistle.com>
PR:		22352